mongodb-azure/device_simulation/requirements.txt
azure-functions
azure.iot.device
azure.iot.hub